refactor: Move DetailViewHolder out of DetailAdapter
The only real content changes are:

1. The companion object
2. Formatting the two constructors to be multiline instead of one looong line

Otherwise it's copy and paste of the DetailViewHolder and the imports that it needs.
